SQUASH_ME: Partial cleanup
