+++ plugin/rust/src/lib.rs
@@ -5,39 +5,4 @@
 // Author: Simon Brummer (simon.brummer@posteo.de)
 // Description: Rust backend library entry point
 
-#[cxx::bridge(namespace = "rust_ffi")]
-mod ffi {
-    extern "Rust" {
-        type Backend;
-
-        fn number(&self) -> i32;
-        fn set_number(&mut self, number: i32);
-        fn increment_number(&mut self);
-
-        fn make_backend() -> Box<Backend>;
-    }
-}
-
-#[derive(Default)]
-pub struct Backend {
-    number: i32,
-}
-
-impl Backend {
-    pub fn number(&self) -> i32 {
-        self.number
-    }
-
-    pub fn set_number(&mut self, number: i32) {
-        self.number = number
-    }
-
-    pub fn increment_number(&mut self) {
-        self.set_number(self.number + 1)
-    }
-}
-
-pub fn make_backend() -> Box<Backend> {
-    Box::<Backend>::default()
-}
-
+pub mod backend;
